+++ packages/react/src/index.tsx
@@ -1,7 +1,7 @@
-import './styles/globals.css'
+import "./styles/globals.css";
 
-export * from './components/box'
-export * from './components/button'
-export * from './components/heading'
-export * from './components/input-text'
-export * from './components/text'
+export * from "./components/box";
+export * from "./components/button";
+export * from "./components/heading";
+export * from "./components/input-text";
+export * from "./components/text";